% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bulk_request.R
\name{ssar_bulk_request}
\alias{ssar_bulk_request}
\title{Request Bulk Report}
\usage{
ssar_bulk_request(
  id = NULL,
  subdomain = Sys.getenv("SSAR_SUBDOMAIN"),
  apikey = Sys.getenv("SSAR_APIKEY")
)
}
\arguments{
\item{id}{Id of the bulk job #required}

\item{subdomain}{The account subdomain}

\item{apikey}{The api key from the account}
}
\value{
list of the bulk keyword ranking report for the requested ID
}
\description{
Retrieve a bulk report using a bulk report id. Intended to be used with the \code{get_bulk_rankings} function but can be used independently as well.
}
\examples{
\dontrun{
ssar_bulk_request(id = {report_id})
}

}
