% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bulk_request.R
\name{ssar_bulk_request}
\alias{ssar_bulk_request}
\title{Request Bulk Request Report}
\usage{
ssar_bulk_request(
  id = NULL,
  subdomain = Sys.getenv("SSAR_SUBDOMAIN"),
  apikey = Sys.getenv("SSAR_APIKEY")
)
}
\arguments{
\item{id}{Id of the bulk job #required}

\item{subdomain}{The account subdomain}

\item{apikey}{The api key from the account}
}
\value{
The table with all keywords ranking information for the requested ID
}
\description{
Retrieve a bulk report of all the rankings or for specific sites.
}
\examples{
\dontrun{
ssar_bulk_request(id = {report_id})
}

}
