% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bulk_request.R
\name{ssar_bulk_request}
\alias{ssar_bulk_request}
\title{Request Bulk Rankings Report}
\usage{
ssar_bulk_request(
  id = NULL,
  subdomain = Sys.getenv("STATR_SUBDOMAIN"),
  apikey = Sys.getenv("STATR_APIKEY")
)
}
\arguments{
\item{id}{Id of the bulk job}

\item{subdomain}{The account subdomain}

\item{apikey}{The api key from the account}
}
\description{
Retrieve a bulk report of all the rankings or for specific sites.
}
