% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rankings.R
\name{ssar_rankings}
\alias{ssar_rankings}
\title{Get Rankings}
\usage{
ssar_rankings(
  keywordid = NULL,
  fromdate = NULL,
  todate = NULL,
  start = 0,
  results = 100,
  subdomain = Sys.getenv("SSAR_SUBDOMAIN"),
  apikey = Sys.getenv("SSAR_APIKEY")
)
}
\arguments{
\item{keywordid}{The keyword id (required)}

\item{fromdate}{Default is 100 most recent results (optional) Format is a string YYYY-MM-DD}

\item{todate}{Default is most recent ranking day (optional) Format is a string YYYY-MM-DD}

\item{start}{The default is 0 (zero indexed). The starting result for paginated requests}

\item{results}{The default is 100}

\item{subdomain}{The account subdomain}

\item{apikey}{The api key from the account}
}
\description{
Retrieve a list of all the sites in a particular project
}
