% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites_ranking_distribution.R
\name{ssar_sites_ranking_dist}
\alias{ssar_sites_ranking_dist}
\title{Get Sites Ranking Distribution}
\usage{
ssar_sites_ranking_dist(
  siteid = NULL,
  from_date = as.character(Sys.Date() - 31),
  to_date = as.character(Sys.Date() - 1),
  subdomain = Sys.getenv("SSAR_SUBDOMAIN"),
  apikey = Sys.getenv("SSAR_APIKEY")
)
}
\arguments{
\item{siteid}{\emph{Required} The site id.}

\item{from_date}{\emph{Required} Character string in the 'YYYY-MM-DD' format. Default is -31 days from today}

\item{to_date}{\emph{Required} Character string in the 'YYYY-MM-DD' format. Default is yesterday.}

\item{subdomain}{The account subdomain}

\item{apikey}{The api key from the account}
}
\value{
The dataframe with Google, GoogleBaseRank, and Bing ranking distributions by date
}
\description{
This request returns all ranking distribution records for Google and Bing for a site with the specified id. The maximum date range can be no greater than 31 days.
}
